[web] Add a component to show the progress
